+++ Solutions/Year2019/Day07/Solution.cs
@@ -17,22 +17,22 @@ namespace AdventOfCode.Solutions.Year2019
         protected override string SolvePartOne() => RunProgramForGivenInput(Input).ToString();
         protected override string SolvePartTwo() => RunProgramForGivenInputWithFeedbackloop(Input).ToString();
 
-        public int RunComputer(string programInput, int input) => _intcodeComptuer.Run(programInput, input);
+        public long RunComputer(string programInput, int input) => _intcodeComptuer.Run(programInput, input);
 
-        public int RunProgramForGivenInput(string programInput)
+        public long RunProgramForGivenInput(string programInput)
         {
             var itemsToPermuate = new List<int> { 0, 1, 2, 3, 4 };
             var permutationList = GetPermutations(itemsToPermuate, itemsToPermuate.Count);
 
             // Keep track of the highest output
-            var highestOutput = 0;
+            var highestOutput = 0L;
 
             // The queue which will be used as input for the intcodecomputer
-            var intCodeInput = new Queue<int>();
+            var intCodeInput = new Queue<long>();
 
             foreach (var permutation in permutationList)
             {
-                var output = 0;
+                var output = 0L;
                 foreach (var ampSetting in permutation)
                 {
                     intCodeInput.Enqueue(ampSetting);
@@ -56,17 +56,17 @@ namespace AdventOfCode.Solutions.Year2019
             return highestOutput;
         }
 
-        public int RunProgramForGivenInputWithFeedbackloop(string input)
+        public long RunProgramForGivenInputWithFeedbackloop(string input)
         {
             var itemsToPermuate = new List<int> { 5, 6, 7, 8, 9 };
             var permutationList = GetPermutations(itemsToPermuate, itemsToPermuate.Count);
 
             // Keep track of the highest output
-            var highestOutput = 0;
+            var highestOutput = 0L;
 
             foreach (var permutation in permutationList)
             {
-                var output = 0;
+                var output = 0L;
 
                 // For each value in the permutation we create a computer and supply the permutation value
                 var intcodeComputers = permutation.Select(p => new StatefulIntcodeComputer(input, p)).ToList();
